[fix] data type on send movement data on attendance_evening

+++ lib/src/core/background/background_task.dart
@@ -136,7 +136,8 @@ class MyTaskHandler extends TaskHandler {
               body: jsonEncode({
                 'da_code': sapID,
                 'mv_date': DateFormat('yyyy-MM-dd').format(DateTime.now()),
-                'time_duration': positionCalculationResult.totalDistance,
+                'time_duration':
+                    positionCalculationResult.totalDistance.toInt(),
                 'distance':
                     positionCalculationResult.totalDuration.inMilliseconds
               }),
